Support groups=all shorthand
Provides for editors to use 'show=all' as shorthand for including markers
for all groups, or for 'all' as an argument to the 'groups' parameter when
loading a static map with Kartotherian (via @kartotherian/snapshot).

If 'all' is present in the comma-separated list, map data for all groups is
requested from the MW API.

If group names prefixed with '-' are present in addition to 'all', the data
for these groups is removed from the response before the data is passed to
the map for display.

Bug: T140083
Change-Id: Ia673c5bb96f60c13ba7abaf17d4b4dc3d30dcee9