Support show=all shorthand, cont.
As seen on test pages[1], using show=all shorthand is not yet working.
Where it appears to be failing is in DataManager.prototype.loadGroups,
which is still retrieving data from the DataStore by specific group IDs
(causing a failure in @kartotherian/snapshot[2] when the nonexistent 'all'
group is requested).  This patch updates it to support the desired
functionality.

[1]
https://en.wikivoyage.org/wiki/User:MHolloway_(WMF)/Sandbox
https://en.wikivoyage.beta.wmflabs.org/wiki/User:Mdholloway

[2] https://github.com/kartotherian/snapshot/blob/424e1b6028206216f08b826cee110a40f9a709af/lib/mapdataLoader.js#L53

Bug: T140083
Change-Id: Idb980a31321505a9cb990192ac332e1761eeb9c3