Relay optional revision ID parameter
When `revid` is present, pass it through to the mapdata api as
`revids`, in lieu of `titles`.  This makes it possible to retrieve
historical versions of map data.

This is a breaking change to this library's public API, so usages
must be updated when migrating to this version.  Bumps the published
version to 0.6.0

Bug: T293844
Change-Id: I31bdeaf5731efb45507d7281121ee08618fcc1f8