Fix comments mentioning non-existing jQuery.debounce()
There are apparently many plug-ins for jQuery that provide such a
feature. But it's not part of jQuery itself.

The only place where we actually use this feature in the Kartographer
extension references mw.util.debounce().

The second place where we use this library is in Kartotherian. The
debounce option is unused there.

Change-Id: I03cf3727a24752cd9f0e34e232badacd35747eb6