Catch errors from each group and log
The current approach of passing `.then( resolve, resolve )` causes
error bubbling to stop at this point, skipping the implicit rethrow.
This was an intentional decision, to allow failed groups to be
returned to the caller along with a ".failed" flag.

This patch doesn't change any behavior, other than to log the error
at "warn" level before discarding it.

Bug: T322616
Change-Id: Ib7bb18aeaccb61c3fb22f7ffa46dcd77fd995665