Drop class hierarchy and simplify promises
Group class hierarchy: I found the various flavors of group to be
confusing and it seemed to break the Liskov substitution principle.
I've replaced with a single class which can express all the same
behaviors.

Group IDs are no longer needed so are omitted from the result.  This
is a breaking change so the version number is bumped.

DataManager.load can now return a mix of successful and failed
external groups, same as loadGroups.

DataStore: Unused on the server, and should be unnecessary from the
client.

Follow-ups:
* Commons ExternalData attribution could be rendered by the client.
Kartotherian snapshot currently ignores this field.
* Simplify the output format, maybe leaving ExternalData as a
child of the containing mapframe group and client calculates the
attribution as above.
* Optimize rendering multiple mapframes on a page.  These can be
bundled as one API call.  Fetching *all* groups on the page is
prudent.

Bug: T322616
Change-Id: Ieab409a19105a8532ae8445d3f42e35cb71b71bb