Expand external data in-place
Major change to how data is returned to library consumers:
previously, ExternalData had been pulled out of groups to the
top-level of the returned list, and now it will be expanded without
changing its position in the groups and GeoJSON hierarchy.

Failures are still pulled up to the top level for observability (and
because there is no concept of a nested failure yet).

Also reintroduces a function now called "loadExternalData", giving
consumers access to the ExternalData expansion separately from
mapdata API fetch.

Bug: T328760
Change-Id: If6cbabd84a03586d1bfd79482dc5ee98c19c6d70